Fix for broken guile detection.
GUILE_PKG is not compatible with GUILE_PROGS and must not be
used together (due to GUILE_EFFECTIVE_VERSION).
To find the C header files and libraries, use pkg-config ...
PKG_CHECK_MODULES is needed only for compiling C code...
But this project has not C/C++ code in it, so ... remove it.