Cache pathway-gene-interactors for 3x speedup!
Cache previous results, so that they are not recomputed again,
if the results are already known. Note that this functin accounts
for about 60% of the total execution time of `gene-pathway-annotation`
so any caching at all is a win. In a test of 681 genes, this offers
a 3x speedup in run time. (On my machine, execution time went from
61801 seconds to 19180 seconds)

Of course, on any one gene, or small handful of genes, there will
be zero speedup; there has to be a cache-hit to see any benefit,
i.e. there have to be multiple genes that share teh same pathways.