Additional simplification of this search.
Please apply pull req #126 first, and verify that it works as expected.
After that pull req, visual inspection shows that all that the code is
is a fetch of all of the inheritance links, and a filter. That can be
done directly, in a much speedier and immediate way, as done here.